Finally merging in pull request 4. expect issues.
I just kept all my new functions even if they aren't compatible with the
changes from the request.
If anything is broken, feel free to fix it and make a new pull request
(which I should be able to take care of
quickly this time now that my git installation seems to be working).
When I have time, I still need to add
compatibility with my (slightly older) lights.